Implemented labels in image view screen
